refac: moved programming logic from routes.js to specific files (OngController)
